fixed process username
