+++ accounts/admin.py
@@ -1,62 +1,13 @@
 # -*- encoding: utf-8 -*-
-from django.contrib import admin
 from django import forms
-from django.contrib.auth.admin import UserAdmin, User
-from django.contrib.auth.forms import ReadOnlyPasswordHashField
+from django.contrib import admin
+from django.contrib.auth.admin import UserAdmin
 from django.utils.translation import ugettext_lazy as _
-from constance import config
 
+from .forms import UserCreationForm, UserChangeForm
 from .models import UserProfile
 
-
-class UserCreationForm(forms.ModelForm):
-    """A form for creating new users. Includes all the required
-    fields, plus a repeated password."""
-    password1 = forms.CharField(label='Password', widget=forms.PasswordInput)
-    password2 = forms.CharField(label='Repeat password',
-                                widget=forms.PasswordInput)
-
-    class Meta:
-        model = User
-        fields = '__all__'
-
-    def clean_password2(self):
-        # Check that the two password entries match
-        password1 = self.cleaned_data.get("password1")
-        password2 = self.cleaned_data.get("password2")
-        if password1 and password2 and password1 != password2:
-            raise forms.ValidationError("Passwords didn't match.")
-        return password2
-
-    def save(self, commit=True):
-        # Save the provided password in hashed format
-        user = super(UserCreationForm, self).save(commit=False)
-        user.set_password(self.cleaned_data["password1"])
-        if commit:
-            user.save()
-        return user
-
-
-class UserChangeForm(forms.ModelForm):
-    """A form for updating users. Includes all the fields on
-    the user, but replaces the password field with admin's
-    password hash display field.
-    """
-    password = ReadOnlyPasswordHashField(
-        label=("Password"),
-        help_text=("Raw passwords are not stored, so there is no way to see "
-                   "this user's password, but you can change the password "
-                   "using <a href=\"password/\">this form</a>."))
-
-    class Meta:
-        model = User
-        fields = '__all__'
-
-    def clean_password(self):
-        # Regardless of what the user provides, return the initial value.
-        # This is done here, rather than on the field, because the
-        # field does not have access to the initial value
-        return self.initial["password"]
+from constance import config
 
 
 class MyUserAdmin(UserAdmin):
@@ -96,7 +47,7 @@ class MyUserAdmin(UserAdmin):
 
         @property
         def short_description(self):
-            return 'Doładuj wybrane konta o %s' % config.ADMIN_TOPUP
+            return u'Doładuj wybrane konta o %s' % config.ADMIN_TOPUP
 
     topup = Topup()
 
@@ -110,7 +61,7 @@ class MyUserAdmin(UserAdmin):
         for user in queryset:
             user.is_vip = True
             user.save(update_fields=['is_vip'])
-    make_vip.short_description = 'Uczyń użytkownikiem VIP'
+    make_vip.short_description = u'Uczyń użytkownikiem VIP'
 
     def block(modeladmin, request, queryset):
         for user in queryset:
