+++ accounts/utils.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+import uuid
 from unidecode import unidecode
